consumer-notifications/requirements.txt
kafka-python
python-dotenv
six
